+++ lib/super/link.rb
@@ -80,6 +80,11 @@ module Super
       }
     end
 
+    def self.polymorphic_parts(*parts_tail)
+      parts_head = Super.configuration.path.strip.gsub(%r{\A/+}, "").gsub(%r{/+\z}, "").strip.split("/")
+      parts_head + parts_tail
+    end
+
     def initialize(text, href, **options)
       @text = text
       @href = href