+++ test/dummy/rails51_sprockets3/config/initializers/super.rb
@@ -1,5 +1,3 @@
 Super.configuration do |c|
   c.title = "My Admin Site"
-  c.controller_namespace = "admin"
-  c.route_namespace = "admin"
 end